Move more common UI files to :core:ui
