+++ app/src/main/java/com/geeksville/mesh/ui/contact/ContactItem.kt
@@ -51,8 +51,8 @@ import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
 import com.geeksville.mesh.AppOnlyProtos
 import com.geeksville.mesh.model.Contact
-import com.geeksville.mesh.ui.common.components.SecurityIcon
 import org.meshtastic.core.strings.R
+import org.meshtastic.core.ui.component.SecurityIcon
 import org.meshtastic.core.ui.theme.AppTheme
 
 @Suppress("LongMethod")
