+++ app/src/main/java/com/geeksville/mesh/ui/settings/radio/components/EditChannelDialog.kt
@@ -45,11 +45,11 @@ import androidx.compose.ui.unit.dp
 import com.geeksville.mesh.ChannelProtos
 import com.geeksville.mesh.channelSettings
 import com.geeksville.mesh.copy
-import com.geeksville.mesh.ui.common.components.EditBase64Preference
-import com.geeksville.mesh.ui.common.components.PositionPrecisionPreference
 import org.meshtastic.core.model.Channel
 import org.meshtastic.core.strings.R
+import org.meshtastic.core.ui.component.EditBase64Preference
 import org.meshtastic.core.ui.component.EditTextPreference
+import org.meshtastic.core.ui.component.PositionPrecisionPreference
 import org.meshtastic.core.ui.component.SwitchPreference
 
 @Suppress("LongMethod")
