+++ app/src/main/java/com/geeksville/mesh/ui/settings/radio/components/SecurityConfigItemList.kt
@@ -46,15 +46,15 @@ import androidx.navigation.NavController
 import com.geeksville.mesh.ConfigProtos.Config.SecurityConfig
 import com.geeksville.mesh.config
 import com.geeksville.mesh.copy
-import com.geeksville.mesh.ui.common.components.EditBase64Preference
-import com.geeksville.mesh.ui.common.components.EditListPreference
 import com.geeksville.mesh.ui.node.NodeActionButton
 import com.geeksville.mesh.ui.settings.radio.RadioConfigViewModel
-import com.geeksville.mesh.util.encodeToString
-import com.geeksville.mesh.util.toByteString
 import com.google.protobuf.ByteString
+import org.meshtastic.core.model.util.encodeToString
+import org.meshtastic.core.model.util.toByteString
 import org.meshtastic.core.strings.R
 import org.meshtastic.core.ui.component.CopyIconButton
+import org.meshtastic.core.ui.component.EditBase64Preference
+import org.meshtastic.core.ui.component.EditListPreference
 import org.meshtastic.core.ui.component.PreferenceCategory
 import org.meshtastic.core.ui.component.SwitchPreference
 import java.security.SecureRandom
