+++ core/model/src/main/kotlin/org/meshtastic/core/model/util/ByteStringExtensions.kt
@@ -15,11 +15,12 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.util
+package org.meshtastic.core.model.util
 
 import android.util.Base64
 import com.google.protobuf.ByteString
 import com.google.protobuf.kotlin.toByteString
 
 fun ByteString.encodeToString() = Base64.encodeToString(this.toByteArray(), Base64.NO_WRAP)
+
 fun String.toByteString() = Base64.decode(this, Base64.NO_WRAP).toByteString()
