+++ core/ui/build.gradle.kts
@@ -26,6 +26,7 @@ android { namespace = "org.meshtastic.core.ui" }
 dependencies {
     implementation(projects.core.database)
     implementation(projects.core.model)
+    implementation(projects.core.navigation)
     implementation(projects.core.prefs)
     implementation(projects.core.proto)
     implementation(projects.core.strings)
