+++ core/ui/detekt-baseline.xml
@@ -3,18 +3,24 @@
   <ManuallySuppressedIssues/>
   <CurrentIssues>
     <ID>ComposableParamOrder:AlertDialogs.kt$SimpleAlertDialog</ID>
+    <ID>ComposableParamOrder:EditBase64Preference.kt$EditBase64Preference</ID>
     <ID>ComposableParamOrder:EditTextPreference.kt$EditTextPreference</ID>
+    <ID>ComposableParamOrder:MainAppBar.kt$MainAppBar</ID>
     <ID>ComposableParamOrder:MaterialBatteryInfo.kt$MaterialBatteryInfo</ID>
     <ID>ComposableParamOrder:NodeChip.kt$NodeChip</ID>
     <ID>ComposableParamOrder:SettingsItem.kt$SettingsItem</ID>
     <ID>ComposableParamOrder:SignalInfo.kt$SignalInfo</ID>
     <ID>ComposableParamOrder:SwitchPreference.kt$SwitchPreference</ID>
     <ID>ContentSlotReused:AdaptiveTwoPane.kt$second</ID>
+    <ID>LambdaParameterEventTrailing:MainAppBar.kt$onClickChip</ID>
     <ID>LongMethod:EditTextPreference.kt$@Composable fun EditTextPreference( title: String, value: String, enabled: Boolean, isError: Boolean, keyboardOptions: KeyboardOptions, keyboardActions: KeyboardActions, onValueChanged: (String) -&gt; Unit, modifier: Modifier = Modifier, summary: String? = null, maxSize: Int = 0, // max_size - 1 (in bytes) onFocusChanged: (FocusState) -&gt; Unit = {}, trailingIcon: (@Composable () -&gt; Unit)? = null, visualTransformation: VisualTransformation = VisualTransformation.None, )</ID>
     <ID>MagicNumber:EditIPv4Preference.kt$0xff</ID>
     <ID>MagicNumber:EditIPv4Preference.kt$16</ID>
     <ID>MagicNumber:EditIPv4Preference.kt$24</ID>
     <ID>MagicNumber:EditIPv4Preference.kt$8</ID>
+    <ID>MagicNumber:EditListPreference.kt$12</ID>
+    <ID>MagicNumber:EditListPreference.kt$12345</ID>
+    <ID>MagicNumber:EditListPreference.kt$67890</ID>
     <ID>MagicNumber:LazyColumnDragAndDropDemo.kt$50</ID>
     <ID>ModifierMissing:AdaptiveTwoPane.kt$AdaptiveTwoPane</ID>
     <ID>ModifierMissing:EmojiPicker.kt$EmojiPicker</ID>
@@ -24,6 +30,8 @@
     <ID>ModifierMissing:LoraSignalIndicator.kt$Rssi</ID>
     <ID>ModifierMissing:LoraSignalIndicator.kt$Snr</ID>
     <ID>ModifierMissing:LoraSignalIndicator.kt$SnrAndRssi</ID>
+    <ID>ModifierMissing:PreferenceDivider.kt$PreferenceDivider</ID>
+    <ID>ModifierMissing:SecurityIcon.kt$SecurityIcon</ID>
     <ID>ModifierMissing:SettingsItem.kt$SettingsItem</ID>
     <ID>ModifierMissing:SettingsItem.kt$SettingsItemDetail</ID>
     <ID>ModifierMissing:SettingsItem.kt$SettingsItemSwitch</ID>
@@ -38,9 +46,12 @@
     <ID>ModifierReused:TextDividerPreference.kt$Row(modifier = modifier.fillMaxWidth().padding(all = 16.dp), verticalAlignment = Alignment.CenterVertically) { Text( text = title, style = MaterialTheme.typography.bodyLarge, color = if (!enabled) { MaterialTheme.colorScheme.onSurface.copy(alpha = 0.38f) } else { Color.Unspecified }, ) if (trailingIcon != null) { Icon(trailingIcon, "trailingIcon", modifier = modifier.fillMaxWidth().wrapContentWidth(Alignment.End)) } }</ID>
     <ID>MultipleEmitters:PreferenceCategory.kt$PreferenceCategory</ID>
     <ID>ParameterNaming:BitwisePreference.kt$onItemSelected</ID>
+    <ID>ParameterNaming:DropDownPreference.kt$onItemSelected</ID>
     <ID>ParameterNaming:EditIPv4Preference.kt$onValueChanged</ID>
+    <ID>ParameterNaming:EditListPreference.kt$onValuesChanged</ID>
     <ID>ParameterNaming:EditPasswordPreference.kt$onValueChanged</ID>
     <ID>ParameterNaming:EditTextPreference.kt$onValueChanged</ID>
+    <ID>ParameterNaming:PositionPrecisionPreference.kt$onValueChanged</ID>
     <ID>ParameterNaming:PreferenceFooter.kt$onNegativeClicked</ID>
     <ID>ParameterNaming:PreferenceFooter.kt$onPositiveClicked</ID>
     <ID>ParameterNaming:SlidingSelector.kt$onOptionSelected</ID>
@@ -50,5 +61,6 @@
     <ID>PreviewPublic:SignalInfo.kt$SignalInfoPreview</ID>
     <ID>PreviewPublic:SignalInfo.kt$SignalInfoSelfPreview</ID>
     <ID>PreviewPublic:SignalInfo.kt$SignalInfoSimplePreview</ID>
+    <ID>UnusedParameter:DropDownPreference.kt$modifier: Modifier = Modifier</ID>
   </CurrentIssues>
 </SmellBaseline>
