+++ core/ui/src/main/kotlin/org/meshtastic/core/ui/component/PreferenceDivider.kt
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.ui.common.components
+package org.meshtastic.core.ui.component
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.HorizontalDivider
